Move docs edits to separate file
